content/events/test/test_bug457672.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=457672
-->
<head>
  <title>Test for Bug 457672</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=457672">Mozilla Bug 457672</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 457672 **/

var windowBlurCount = 0;

function setUserPref(reset) {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  const prefSvcContractID = "@mozilla.org/preferences-service;1";
  const prefSvcIID = Components.interfaces.nsIPrefService;
  var prefs = Components.classes[prefSvcContractID].getService(prefSvcIID)
                                                   .getBranch("browser.link.");
  if (reset) {
    prefs.clearUserPref("open_newwindow");
  } else {
    prefs.setIntPref("open_newwindow", 3);
  }
}

function listener(evt) {
  if (evt.type == "focus") {
    is(windowBlurCount, 1,
       "Window should have got blur event when opening a new tab!");
    setUserPref(true);
    document.getElementsByTagName("a")[0].focus();
    SimpleTest.finish();
  } else if (evt.type == "blur") {
    ++windowBlurCount;
  }
  document.getElementById('log').textContent += evt.target + ":" + evt.type + "\n";
}

function startTest() {
  setUserPref(false);
  document.getElementsByTagName("a")[0].focus();
  // Note, focus/blur don't bubble
  window.addEventListener("focus", listener, false);
  window.addEventListener("blur", listener, false);
  window.open("bug457672.html", "", "");
}

addLoadEvent(startTest);
SimpleTest.waitForExplicitFinish();

</script>
</pre>
<pre id="log">
</pre>
</body>
</html>
